% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_criteria_direction}
\alias{check_criteria_direction}
\title{Sprawdzenie poprawności kierunku kryteriów}
\usage{
check_criteria_direction(crit_names)
}
\arguments{
\item{crit_names}{nazwy kolumn kryteriów}
}
\description{
Funkcja pomocnicza weryfikująca, czy nazwy w CRITERIA_DIRECTION
zgadzają się z kolumnami macierzy decyzyjnej.
}
\keyword{internal}
