% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{CRITERIA_DIRECTION}
\alias{CRITERIA_DIRECTION}
\title{Kierunek optymalizacji dla poszczególnych kryteriów}
\format{
Nazwany wektor znakowy o długości 8
\describe{
  \item{severity}{poziom powagi incydentu – cost}
  \item{malware}{wskaźnik obecności malware – cost}
  \item{anomaly}{wynik anomalii – cost}
  \item{packet_load}{średnia długość pakietów – cost}
  \item{penetration}{głębokość penetracji sieci – cost}
  \item{security_det}{liczba sygnałów bezpieczeństwa – benefit}
  \item{asset_crit}{krytyczność zasobu (długość opisu) – benefit}
  \item{geo_risk}{pochodzenie z kraju wysokiego ryzyka – cost}
}
}
\usage{
CRITERIA_DIRECTION
}
\description{
Stała definiująca, czy dane kryterium jest typu "cost" (im mniejsza wartość,
tym lepiej) czy "benefit" (im większa wartość, tym lepiej).
Wykorzystywana do poprawnego przetwarzania macierzy decyzyjnej w metodach
TOPSIS, Fuzzy TOPSIS oraz VIKOR.
}
\keyword{datasets}
