% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{cyber}
\alias{cyber}
\title{Szybki meta-ranking cyberbezpieczeństwa}
\usage{
cyber(wagi = "bwm", ...)
}
\arguments{
\item{wagi}{"entropia" lub "bwm" (domyślnie "bwm")}

\item{...}{dodatkowe argumenty przekazywane do rozmyty_meta_ranking}
}
\value{
Obiekt meta-rankingu (niewidoczny), ale od razu drukuje tabelkę
}
\description{
Jedno słowo → pełny meta-ranking z domyślnymi danymi i wagami BWM
}
