% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdm_bwm.R
\name{cyber_bwm_weights}
\alias{cyber_bwm_weights}
\title{Wagi BWM – ręcznie ustawione na razie bo nie wychodzi mi inaczej}
\usage{
cyber_bwm_weights()
}
\description{
Wagi BWM – ręcznie ustawione na razie bo nie wychodzi mi inaczej
}
