% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyberrank_main.R
\name{cyber_rank_all}
\alias{cyber_rank_all}
\title{Porównanie wyników trzech metod MCDA}
\usage{
cyber_rank_all()
}
\value{
lista z wynikami trzech metod oraz wydrukowana tabela porównawcza
}
\description{
Porównanie wyników trzech metod MCDA
}
