% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdm_fuzzy_vikor.R
\name{cyber_rank_fuzzy_vikor}
\alias{cyber_rank_fuzzy_vikor}
\title{Fuzzy VIKOR (własna impl. z defuzz)}
\usage{
cyber_rank_fuzzy_vikor(v = 0.5)
}
\arguments{
\item{v}{parametr strategii (0.5 = równowaga)}
}
\description{
Fuzzy VIKOR (własna impl. z defuzz)
}
