% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyberrank_classic.R, R/mcdm_fuzzy_topsis.R
\name{cyber_rank_topsis}
\alias{cyber_rank_topsis}
\title{Ranking BWM + TOPSIS (klasyczny)}
\usage{
cyber_rank_topsis()

cyber_rank_topsis()
}
\value{
data.frame z rankingiem typów ataków według metody TOPSIS

data.frame z rankingiem typów ataków
}
\description{
Metoda stosuje rozmycie trójkątne ±20% wokół wartości ostrych,
a następnie defuzzyfikację (centrum ciężkości) i klasyczny TOPSIS.
}
