% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria_calculation.R
\name{prepare_cyber_criteria}
\alias{prepare_cyber_criteria}
\title{Przygotowanie macierzy kryteriów dla wszystkich metod}
\usage{
prepare_cyber_criteria(min_attacks = 20)
}
\arguments{
\item{min_attacks}{minimalna liczba incydentów wymagana do uwzględnienia typu ataku (domyślnie 20)}
}
\value{
data.frame z zagregowanymi kryteriami dla każdego typu ataku
}
\description{
Przygotowanie macierzy kryteriów dla wszystkich metod
}
