% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\name{cyber_expert_panel}
\alias{cyber_expert_panel}
\title{Zagregowane dane o incydentach cyberbezpieczeństwa do analizy MCDA}
\format{
Ramka danych (data frame) o 60 wierszach i 7 kolumnach:
\describe{
  \item{ExpertID}{Identyfikator eksperta (Geo-location Data)}
  \item{Alternative}{Typ ataku cybernetycznego (DDoS, Malware, Intrusion)}
  \item{crit_severity}{Dotkliwość ataku (skala 1-9)}
  \item{crit_anomaly}{Średni poziom anomalii (wartość ciągła)}
  \item{crit_complexity}{Techniczna złożoność ataku / długość pakietu}
  \item{crit_impact}{Głębokość penetracji sieci (skala 1-9)}
  \item{crit_stealth}{Trudność wykrycia ataku (skala odwrócona, 1-9)}
}
}
\usage{
data(cyber_expert_panel)
}
\description{
Zbiór danych zawierający zagregowane informacje o incydentach
cyberbezpieczeństwa, pogrupowane według typu ataku.
Dane zostały przygotowane do wykorzystania w metodach
wielokryterialnego wspomagania decyzji (MCDA), takich jak
TOPSIS, Fuzzy TOPSIS czy PROMETHEE.
}
\details{
Każdy wiersz reprezentuje jedną alternatywę decyzyjną zdefiniowaną jako kombinacja kraju, typu ataku oraz scenariusza (3 alternatywy na każdą parę kraj × typ ataku).
Kolumny opisują średnie wartości kryteriów obliczone na podstawie
co najmniej 20 zarejestrowanych incydentów.
}
