% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdm_fuzzy_topsis.R
\name{cyber_rank_fuzzy_topsis}
\alias{cyber_rank_fuzzy_topsis}
\title{Fuzzy TOPSIS (własna impl. z defuzz dla prostoty)}
\usage{
cyber_rank_fuzzy_topsis()
}
\description{
Fuzzy TOPSIS (własna impl. z defuzz dla prostoty)
}
