% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyberrank_classic.R
\name{cyber_rank_topsis}
\alias{cyber_rank_topsis}
\title{Ranking BWM + TOPSIS (klasyczny)}
\usage{
cyber_rank_topsis()
}
\value{
data.frame z rankingiem typów ataków według metody TOPSIS
}
\description{
Ranking BWM + TOPSIS (klasyczny)
}
