% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria_calculation.R
\name{prepare_cyber_criteria}
\alias{prepare_cyber_criteria}
\title{Przygotowanie macierzy kryteriów dla wszystkich metod}
\usage{
prepare_cyber_criteria(min_attacks = 20)
}
\description{
Przygotowanie macierzy kryteriów dla wszystkich metod
}
